rename and test PagesHeader
